Refactored WordsLayout and deleted wordsFilteredByCategory
