+++ frontend/app/assets/javascripts/views/pages/index.js
@@ -50,7 +50,6 @@ import PageDialectViewAlphabet from './explore/dialect/learn/alphabet/'
 import PageDialectViewCharacter from './explore/dialect/learn/alphabet/view'
 import PageDialectLearnWordsCategories from './explore/dialect/learn/words/categories'
 import PageDialectLearnPhrasesCategories from './explore/dialect/learn/phrases/categories'
-import PageDialectLearnWordsFilteredByCategory from './explore/dialect/learn/words/wordsFilteredByCategory'
 import PageDialectLearnPhrasesFilteredByCategory from './explore/dialect/learn/phrases/phrasesFilteredByCategory'
 
 import PageDialectImmersionList from './explore/dialect/immersion'
@@ -138,7 +137,6 @@ export {
   PageDialectLearn,
   PageDialectMedia,
   PageDialectLearnWords,
-  PageDialectLearnWordsFilteredByCategory,
   PageDialectLearnWordsCategories,
   PageDialectLearnPhrases,
   PageDialectLearnPhrasesFilteredByCategory,
